Refactored codebase to use "Coordinates" and put files in pertinent subpackages.
