+++ src/main/java/mars/algorithm/Algorithm.java
@@ -1,4 +1,8 @@
-package mars;
+package mars.algorithm;
+
+import mars.rover.MarsRover;
+import mars.out.Output;
+import mars.map.TerrainMap;
 
 /**
  * Abstract class from which all path-finding algorithms inherit.
