+++ src/main/java/mars/ui/UserInterface.java
@@ -1,6 +1,10 @@
-package mars;
+package mars.ui;
 
 //import com.sun.org.apache.xml.internal.security.algorithms.JCEMapper;
+
+import mars.algorithm.Algorithm;
+import mars.out.Output;
+import mars.rover.Rover;
 
 /**
  * Abstract class from which all user interfaces inherit.
